* Cleaned concept of MediaItem

+++ jolivia.protocol/src/main/java/org/dyndns/jkiddo/dmap/chunks/media/UTCTime.java
@@ -9,7 +9,7 @@ public class UTCTime extends DateChunk
 		this(0);
 	}
 
-	public UTCTime(int i)
+	public UTCTime(long i)
 	{
 		super("mstc", "dmap.utctime", i);
 	}
